chore: split saved gestures into multiple files
