+++ app/common/renderer/components/SessionInspector/GesturesTab/SavedGestureActionsCell.jsx
@@ -1,0 +1,74 @@
+import {IconEdit, IconFileExport, IconPlayerPlay, IconTrash} from '@tabler/icons-react';
+import {Button, Popconfirm, Space, Tooltip} from 'antd';
+import _ from 'lodash';
+import {useTranslation} from 'react-i18next';
+
+import {SCREENSHOT_INTERACTION_MODE} from '../../../constants/screenshot.js';
+
+/**
+ * Final cell of each saved gestures row, containing gesture actions.
+ */
+const SavedGestureActionsCell = (props) => {
+  const {
+    gesture,
+    gesturePointers,
+    showGestureEditor,
+    removeGestureDisplay,
+    exportSavedGesture,
+    setLoadedGesture,
+    deleteSavedGesture,
+    applyClientMethod,
+  } = props;
+  const {t} = useTranslation();
+
+  const playGesture = () => {
+    const actions = {};
+    for (const pointer of gesturePointers) {
+      actions[pointer.name] = pointer.ticks.map((tick) => _.omit(tick, 'id'));
+    }
+    applyClientMethod({methodName: SCREENSHOT_INTERACTION_MODE.GESTURE, args: [actions]});
+  };
+
+  const loadGesture = () => {
+    setLoadedGesture(gesture);
+    showGestureEditor();
+  };
+
+  const deleteGesture = () => {
+    removeGestureDisplay();
+    deleteSavedGesture(gesture.id);
+  };
+
+  return (
+    <Space.Compact>
+      <Tooltip zIndex={3} title={t('Play')}>
+        <Button
+          key="play"
+          type="primary"
+          icon={<IconPlayerPlay size={18} />}
+          onClick={() => playGesture()}
+        />
+      </Tooltip>
+      <Tooltip zIndex={3} title={t('Edit')}>
+        <Button icon={<IconEdit size={18} />} onClick={() => loadGesture()} />
+      </Tooltip>
+      <Tooltip zIndex={3} title={t('Export to File')}>
+        <Button icon={<IconFileExport size={18} />} onClick={() => exportSavedGesture(gesture)} />
+      </Tooltip>
+      <Tooltip zIndex={3} title={t('Delete')}>
+        <Popconfirm
+          zIndex={4}
+          title={t('confirmDeletion')}
+          placement="topRight"
+          okText={t('OK')}
+          cancelText={t('Cancel')}
+          onConfirm={() => deleteGesture()}
+        >
+          <Button icon={<IconTrash size={18} />} />
+        </Popconfirm>
+      </Tooltip>
+    </Space.Compact>
+  );
+};
+
+export default SavedGestureActionsCell;
